refactor: 색상 수정 요청 로직 재수정
